refactor: extraer helper de upload de imagen a src/lib/imageUpload.js
El mismo flujo (canvas resize a 1200px + POST a /upload + parse response)
estaba duplicado en 3 archivos:
- pages/dashboard/home-config.jsx
- pages/dashboard/postres.jsx
- src/components/Resenas.jsx

Nuevo módulo `src/lib/imageUpload.js` exporta:
- `redimensionarImagen(file)` → File PNG redimensionado a max 1200 px.
- `subirImagen(file, apiBase, token)` → hace resize + POST a /upload
  y devuelve `{ fileUrl, fileName }`.

Los 3 archivos ahora hacen `import { subirImagen } from "@/src/lib/imageUpload"`
y reemplazan ~30-40 líneas de inline boilerplate por una llamada de 1
línea. ~100 líneas eliminadas en total.

Sin cambios funcionales: misma URL final, mismo MAX_LADO_PX=1200,
mismas validaciones. Solo deduplicación.
